+++ reddit_api_kernel/mod_action.ml
@@ -1,7 +1,7 @@
 open! Core
 
 module Id = struct
-  type t = Uuid.Unstable.t [@@deriving sexp]
+  type t = Uuid.Unstable.t [@@deriving compare, hash, sexp]
 
   let of_uuid = Fn.id
   let to_uuid = Fn.id
@@ -23,8 +23,13 @@ let id = required_field "id" (string >> Id.of_json_string)
 let action = required_field "action" string
 let details = required_field "details" string
 let created = required_field "created_utc" time_sec_since_epoch
+
+let target_author =
+  required_field "target_author" (string >> Username.of_string_or_deleted)
+;;
+
 let target_title = required_field "target_title" string
 let target_fullname = required_field "target_fullname" (string >> Thing.Fullname.of_string)
 let target_permalink = required_field "target_permalink" uri
 let subreddit_name = required_field "subreddit_name" subreddit_name
-let moderator = required_field "moderator" (string >> Username.of_string_or_deleted)
+let moderator = required_field "mod" (string >> Username.of_string_or_deleted)